+++ backend/src/tests/Unit/Repository/CorporationMemberRepositoryTest.php
@@ -2,15 +2,19 @@
 
 namespace Tests\Unit\Repository;
 
+use Doctrine\ORM\EntityManagerInterface;
+use Neucore\Entity\Character;
 use Neucore\Entity\Corporation;
 use Neucore\Entity\CorporationMember;
+use Neucore\Entity\Player;
 use Neucore\Factory\RepositoryFactory;
+use PHPUnit\Framework\TestCase;
 use Tests\Helper;
 
-class CorporationMemberRepositoryTest extends \PHPUnit\Framework\TestCase
+class CorporationMemberRepositoryTest extends TestCase
 {
     /**
-     * @var \Doctrine\ORM\EntityManagerInterface
+     * @var EntityManagerInterface
      */
     private static $em;
 
@@ -23,10 +27,12 @@ class CorporationMemberRepositoryTest extends \PHPUnit\Framework\TestCase
         $helper->emptyDb();
         self::$em = $helper->getEm();
 
+        $player1 = (new Player())->setName('Player 1');
+        $char1 = (new Character())->setId(1)->setName('Char 1')->setPlayer($player1);
         $corp1 = (new Corporation())->setId(1)->setName('Corp 1')->setTicker('C1');
         $corp2 = (new Corporation())->setId(2)->setName('Corp 2')->setTicker('C2');
         $member1 = (new CorporationMember())->setId(10)->setName('Member 1')->setCorporation($corp1)
-            ->setLogonDate(new \DateTime('now -112 days +1 hour'));
+            ->setLogonDate(new \DateTime('now -112 days +1 hour'))->setCharacter($char1);
         $member2 = (new CorporationMember())->setId(20)->setName('Member 2')->setCorporation($corp1)
             ->setLogonDate(new \DateTime('now -111 days +1 hour'));
         $member3 = (new CorporationMember())->setId(30)->setName('Member 3')->setCorporation($corp1)
@@ -34,6 +40,8 @@ class CorporationMemberRepositoryTest extends \PHPUnit\Framework\TestCase
         $member4 = (new CorporationMember())->setId(40)->setName('Member 4')->setCorporation($corp2)
             ->setLogonDate(new \DateTime('now -111 days +1 hour'));
 
+        self::$em->persist($player1);
+        self::$em->persist($char1);
         self::$em->persist($corp1);
         self::$em->persist($corp2);
         self::$em->persist($member1);
@@ -44,13 +52,27 @@ class CorporationMemberRepositoryTest extends \PHPUnit\Framework\TestCase
         self::$em->flush();
     }
 
-    public function testFindByLogonDateActive()
+    public function testResetCriteria()
     {
         $repository = (new RepositoryFactory(self::$em))->getCorporationMemberRepository();
 
-        $actual1 = $repository->findByLogonDate(1, null, 109);
-        $actual2 = $repository->findByLogonDate(1, null, 110);
-        $actual3 = $repository->findByLogonDate(1, null, 112);
+        $repository->setAccount(false);
+        $repository->setActive(110);
+        $repository->setInactive(110);
+        $this->assertSame(0, count($repository->findMatching(1)));
+
+        $repository->resetCriteria();
+
+        $this->assertSame(3, count($repository->findMatching(1)));
+    }
+
+    public function testFindMatchingActive()
+    {
+        $repository = (new RepositoryFactory(self::$em))->getCorporationMemberRepository();
+
+        $actual1 = $repository->setActive(109)->findMatching(1);
+        $actual2 = $repository->setActive(110)->findMatching(1);
+        $actual3 = $repository->setActive(112)->findMatching(1);
 
         $this->assertSame(0, count($actual1));
         $this->assertSame(1, count($actual2));
@@ -65,13 +87,13 @@ class CorporationMemberRepositoryTest extends \PHPUnit\Framework\TestCase
         $this->assertSame('Corp 1', $actual3[1]->getCorporation()->getName());
     }
 
-    public function testFindByLogonDateInactive()
+    public function testFindMatchingInactive()
     {
         $repository = (new RepositoryFactory(self::$em))->getCorporationMemberRepository();
 
-        $actual1 = $repository->findByLogonDate(1, 109);
-        $actual2 = $repository->findByLogonDate(1, 110);
-        $actual3 = $repository->findByLogonDate(1, 112);
+        $actual1 = $repository->setInactive(109)->findMatching(1);
+        $actual2 = $repository->setInactive(110)->findMatching(1);
+        $actual3 = $repository->setInactive(112)->findMatching(1);
 
         $this->assertSame(3, count($actual1));
         $this->assertSame(2, count($actual2));
@@ -87,15 +109,31 @@ class CorporationMemberRepositoryTest extends \PHPUnit\Framework\TestCase
         $this->assertSame('Corp 1', $actual2[0]->getCorporation()->getName());
     }
 
-    public function testFindByLogonDateActiveRange()
+    public function testFindMatchingActiveRange()
     {
         $repository = (new RepositoryFactory(self::$em))->getCorporationMemberRepository();
 
-        $actual = $repository->findByLogonDate(1, 110, 111);
+        $actual = $repository->setInactive(110)->setActive(111)->findMatching(1);
 
         $this->assertSame(1, count($actual));
 
         $this->assertSame('Member 2', $actual[0]->getName());
         $this->assertSame('Corp 1', $actual[0]->getCorporation()->getName());
+    }
+
+    public function testFindMatchingWithOutAccount()
+    {
+        $repository = (new RepositoryFactory(self::$em))->getCorporationMemberRepository();
+
+        $actual1 = $repository->setAccount(true)->findMatching(1);
+        $actual2 = $repository->setAccount(false)->findMatching(1);
+
+        $this->assertSame(1, count($actual1));
+        $this->assertSame(2, count($actual2));
+
+        $this->assertSame('Member 1', $actual1[0]->getName());
+        $this->assertSame('Char 1', $actual1[0]->getCharacter()->getName());
+        $this->assertSame('Member 3', $actual2[0]->getName());
+        $this->assertSame('Member 2', $actual2[1]->getName());
     }
 }